v0.9.7

+++ nftfw/nf_edit_dbfns.py
@@ -20,6 +20,7 @@ class DbFns:
         self.cf = cf
         self.port = None
         self.pattern = None
+        self.matches = None
 
     def execute(self):
         """Scheduler access point
@@ -45,11 +46,12 @@ class DbFns:
         fn = getattr(self, cmd)
         self.port = self.cf.editargs['port']
         self.pattern = self.cf.editargs['pattern']
-
+        self.matches = self.cf.editargs['matches']
         files = fn(self.cf.editargs['iplist'])
         if cmd == 'add':
             return 0
         return files
+
 
     def add(self, iplist):
         """Add list of ip addresses to the database
@@ -60,39 +62,18 @@ class DbFns:
         Parameters
         ----------
         iplist : List[str]
+
+        Returns
+        -------
+        int: Number of files added
+
         """
 
-        # create a worklist for blacklist
-        template = {
-            'ports': self.port,
-            'pattern': self.pattern,
-            'matchcount': 1,
-            'incidents': 1
-            }
+        return self.blacklist(iplist,
+                              create_files=False,
+                              set_min_matches=False)
 
-        normalise = NormaliseAddress(self.cf, 'nftfwedit add')
-        iplist = self.normalise_ip_list(normalise, iplist)
-        if iplist is None:
-            return
-
-        # use blacklist.py methods
-        bl = BlackList(self.cf)
-        fwdb = FwDb(self.cf)
-
-        # check and convert IP addresses
-        for ip in iplist:
-            # check if the ip is in the database
-            indb = fwdb.lookup_by_ip(ip)
-            if any(indb):
-                log.error('%s is already in the database', ip)
-            else:
-                # add this ip
-                bl.db_store(fwdb, ip, template, forcenew=True)
-                log.info('%s added to database', ip)
-
-        fwdb.close()
-
-    def blacklist(self, iplist):
+    def blacklist(self, iplist, create_files=True, set_min_matches=True):
         """Add a set of blacklist directory as .auto
 
         If necessary, add the ip the blacklist database by using add
@@ -101,30 +82,44 @@ class DbFns:
         Makes use of methods from blacklist.py to
         do the hard stuff
 
+        This code is reused by the add action, but with
+        create_files = False
+        which is used to set the value in the blacklist
+        code that prevents file creation.
+
+        setting set_min_matches to False prevents setting
+        of matchcount to a minimum of block_after value
+
         Return number of files added
         """
 
-        normalise = NormaliseAddress(self.cf, 'nftfwedit blacklist')
-        iplist = self.normalise_ip_list(normalise, iplist)
+        # pylint: disable=too-many-locals, too-many-branches
+
+        iplist = self.ensure_legal_addresses(iplist, 'nftfwedit blacklist')
         if iplist is None:
             return 0
 
         fwdb = FwDb(self.cf)
 
         bl = BlackList(self.cf)
+        bl.report_whitelisting = True
+        # Inhibit file creation if wanted
+        if not create_files:
+            bl.create_files = False
 
         # Generate database entries for update
         # use work to retain database values
-        work = {}
+        # check if we need ports and patterns
         toadd = []
+        current = []
         for ip in iplist:
             # check if the ip is in the database
             indb = fwdb.lookup_by_ip(ip)
             if any(indb):
-                # will only be one
-                work[ip] = indb[0]
+                current.append(ip)
             else:
                 toadd.append(ip)
+        fwdb.close()
 
         # any ips need adding?
         if any(toadd):
@@ -132,43 +127,45 @@ class DbFns:
             if self.port is None \
                or self.pattern is None:
                 ipe = ", ".join(toadd)
-                log.error('Pattern or port needed for %s', ipe)
-                fwdb.close()
+                log.error('Pattern and port needed for %s', ipe)
                 return 0
 
-            # dbstore template
-            template = {
+        work = {}
+        for ip in iplist:
+            if ip in current:
+                if self.port is None:
+                    self.port = 'update'
+                if self.pattern is None:
+                    self.pattern = current[ip]['pattern']
+            work[ip] = {
                 'ports': self.port,
                 'pattern': self.pattern,
-                'matchcount': 1,
                 'incidents': 1
-            }
-            for ip in toadd:
-                current = bl.db_store(fwdb, ip, template, forcenew=True)[0]
-                work[ip] = current
-                log.info('%s added to database', ip)
+                }
+            # set matchcount
+            work[ip]['matchcount'] = self.matches
+            min_count = bl.block_after
+            if ip in toadd and set_min_matches:
+                # for new entries start matchcount at the threshold
+                if self.matches < min_count:
+                    work[ip]['matchcount'] = min_count
 
-        # create files
+        # do one call to install_ips to minimise whitelist checking
         added = 0
-
-        for ip in iplist:
-            current = work[ip]
-            # fake matchcount
-            current['matchcount'] = int(bl.block_after)
-            added += bl.install_file(fwdb, current, True)
-            log.info('%s added to blacklist.d', ip)
-
-        fwdb.close()
+        filesinstalled, _ = bl.install_ips(work)
+        if filesinstalled > 0:
+            added += filesinstalled
         return added
 
-    def delete(self, iplist):
+    def delete(self, iplist, delete_from_db = True):
         """Delete a list of ipaddresses address from the database
 
-        Also deletes from the blacklist.d directory if there
+        Also deletes from the blacklist.d directory if there is a file to be deleted
 
         Parameters
         ----------
         iplist : List[str]
+        delete_from_db - delete the database entry if true
 
         Returns
         -------
@@ -176,8 +173,7 @@ class DbFns:
             Number of files deleted
         """
 
-        normalise = NormaliseAddress(self.cf, 'nftfwedit delete')
-        iplist = self.normalise_ip_list(normalise, iplist)
+        iplist = self.ensure_legal_addresses(iplist, 'nftfwedit delete')
         if iplist is None:
             return 0
 
@@ -190,7 +186,7 @@ class DbFns:
         filesdeleted = 0
         for ip in iplist:
             indb = fwdb.lookup_by_ip(ip)
-            if any(indb):
+            if delete_from_db and any(indb):
                 fwdb.delete_ip(ip)
                 ipsdeleted += 1
 
@@ -210,6 +206,12 @@ class DbFns:
 
         return filesdeleted
 
+    def remove(self, iplist):
+        """ Remove a file from the blacklist.d directory
+        leaving the database alone """
+
+        return self.delete(iplist, delete_from_db = False)
+
     @staticmethod
     def normalise_ip_list(normalise, iplist):
         """Normalise and check ip address list
@@ -224,8 +226,7 @@ class DbFns:
         Returns
         -------
         List[str]
-            Return list of addresses
-            None on any errors
+            Return list of valid addresses
         """
 
         added = []
@@ -233,7 +234,20 @@ class DbFns:
             newip = normalise.normal(ip)
             if newip is not None:
                 added.append(newip)
-
-        if len(added) != len(iplist):
-            return None
         return added
+
+    def ensure_legal_addresses(self, iplist, fromwhere):
+        """Normalise a list of addresses and complain
+        when some are not normal
+
+        Return a legal list
+        """
+
+        normalise = NormaliseAddress(self.cf, fromwhere)
+        normal_iplist = self.normalise_ip_list(normalise, iplist)
+        if len(normal_iplist) != len(iplist):
+            deleted = list(set(iplist) - set(normal_iplist))
+            if any(deleted):
+                log.error('Does not pass validation tests: %s',
+                          ", ".join(deleted))
+        return normal_iplist
